Refactor LCD and buzzer code into separate modules
