+++ firmware/dripito-v1/InfusionBA/InfusionBA/Core/Inc/buzzer.h
@@ -1,0 +1,9 @@
+#ifndef BUZZER_H
+#define BUZZER_H
+
+#include "main.h"
+#include <stdint.h>
+
+void Buzzer_PlayFreq(uint16_t freq, uint16_t duration_ms);
+
+#endif // BUZZER_H
